Use iterators instead of generators
Generators are not really needed right now and Iterators are more
compatible than generators. Generators were only introduced in
PHP 5.5 and they work different internally in HHVM making them hard
to thest. Also in the current readers bundled with Plum they are,
as mentioned, not needed and using generators doesn't offer any
advantages.

However, in contrast to the very early version the ReaderInterface
still extends IteratorAggregate (instead of Iterator) which results
in siginificantly less code, since we can use, for example, the
ArrayIterator for ArrayReader