ReaderFactory supports accepts callback
accepts() is no longer part of ReaderInterface, instead ReaderFactory
first calls the accepts callback passed to addReader() (if it exists)
and then calls the ::accepts() method of the reader (if it not exists).
A reader without a ::accepts() method will never be created by
ReaderFactory (except when added with a accepts cb).